requirements.txt
Django==3.2.13
django-environ==0.4.5
djangorestframework==3.12.4
django-cors-headers~=3.1.1
drf-jwt==1.15.1
pyjwt~=1.7.1
requests==2.27.1
django-storages==1.11.1
# celery==5.2.2
redis==3.5.3
django-filter==21.1
drf_nested_forms==1.1.6
Faker==9.5.0
djangorestframework-simplejwt==5.2.0
# for scrapy
attrs==21.2.0
Automat==20.2.0
cffi==1.14.6
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
Pillow==8.3.1
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
queuelib==1.6.1
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
Twisted==21.7.0
typing-extensions==3.10.0.0
w3lib==1.22.0
zope.interface==5.4.0
scrapy-splash==0.9.0
selenium==4.12.0
ipython==8.15.0
